requirements.txt
kytea==0.1.7
regex==2022.10.31
